Shared UI Tests Setup and Tear Down
