+++ WemotelyUITests/EditDashboardTests.swift
@@ -1,27 +1,6 @@
 import XCTest
 
-class EditDashboardTests: XCTestCase {
-    var app: XCUIApplication!
-
-    override func setUp() {
-        super.setUp()
-
-        continueAfterFailure = false
-
-        app = XCUIApplication()
-        app.resetOrientation()
-
-        // Enable it to reset its state
-        app.launchArguments.append("--uitesting")
-    }
-
-    override func tearDown() {
-        super.tearDown()
-        app.resetOrientation()
-    }
-
-    // MARK: - Tests
-
+class EditDashboardTests: BaseUITestCase {
     func testAccountActivation() {
         app.launch()
