+++ WemotelyUITests/JobTests.swift
@@ -1,30 +1,9 @@
 import XCTest
 
-class JobTests: XCTestCase {
-    var app: XCUIApplication!
-
+class JobTests: BaseUITestCase {
     let accountName = "All Inboxes"
     let companyName = "Car Next Door"
     let jobTitle = "Front-End Engineer - CND Growth Team"
-
-    override func setUp() {
-        super.setUp()
-
-        continueAfterFailure = false
-
-        app = XCUIApplication()
-        app.resetOrientation()
-
-        // Enable it to reset its state
-        app.launchArguments.append("--uitesting")
-    }
-
-    override func tearDown() {
-        super.tearDown()
-        app.resetOrientation()
-    }
-
-    // MARK: - Tests
 
     func testNavigationBarTitlePrompt() {
         app.launch()
