+++ WemotelyUITests/JobsSwipeActionTests.swift
@@ -1,28 +1,7 @@
 import XCTest
 
-class JobsSwipeActionTests: XCTestCase {
-    var app: XCUIApplication!
-
+class JobsSwipeActionTests: BaseUITestCase {
     let jobTitle = "Front-End Engineer - CND Growth Team"
-
-    override func setUp() {
-        super.setUp()
-
-        continueAfterFailure = false
-
-        app = XCUIApplication()
-        app.resetOrientation()
-
-        // Enable it to reset its state
-        app.launchArguments.append("--uitesting")
-    }
-
-    override func tearDown() {
-        super.tearDown()
-        app.resetOrientation()
-    }
-
-    // MARK: - Tests
 
     func testSwipeToDelete() {
         let fromAccount = (name: "All Inboxes", index: 0)
