+++ WemotelyUITests/SegueUITests.swift
@@ -1,27 +1,6 @@
 import XCTest
 
-class SegueUITests: XCTestCase {
-    var app: XCUIApplication!
-
-    override func setUp() {
-        super.setUp()
-
-        continueAfterFailure = false
-
-        app = XCUIApplication()
-        app.resetOrientation()
-
-        // Enable it to reset its state
-        app.launchArguments.append("--uitesting")
-    }
-
-    override func tearDown() {
-        super.tearDown()
-        app.resetOrientation()
-    }
-
-    // MARK: - Tests
-
+class SegueUITests: BaseUITestCase {
     func testDetailViewExpandCollapse() {
         app.launch()
