+++ WemotelyUITests/Support/BaseUITestCase.swift
@@ -1,0 +1,22 @@
+import XCTest
+
+class BaseUITestCase: XCTestCase {
+    var app: XCUIApplication!
+
+    override func setUp() {
+        super.setUp()
+
+        continueAfterFailure = false
+
+        app = XCUIApplication()
+        app.resetOrientation()
+
+        // Enable it to reset its state
+        app.launchArguments.append("--uitesting")
+    }
+
+    override func tearDown() {
+        super.tearDown()
+        app.resetOrientation()
+    }
+}
