+++ WemotelyUITests/Support/XCUIApplication+Device.swift
