+++ WemotelyUITests/Support/XCUIApplication+Helper.swift
